Edited ProjectManager and ProjectManagerTest to account for some small modifications
