--- PMSCore.Test/UserMock.cs
@@ -1,8 +1,0 @@
-namespace PMSCore.Test;
-
-/// <summary>
-/// Test User Mock class that implements User interface for dependency injection
-/// </summary>
-internal class UserMock : IUser
-{
-}
